fix(content): dispatch MonetizationEvent instead of CustomEvent
