--- src/content/static/index.ts
@@ -1,13 +1,0 @@
-import { wm2Polyfill } from './polyfill'
-
-function inject(configure: (_script: HTMLScriptElement) => void) {
-  const script = document.createElement('script')
-  configure(script)
-  document.documentElement.appendChild(script)
-  document.documentElement.removeChild(script)
-}
-
-// eslint-disable-next-line @typescript-eslint/no-extra-semi
-;(function injectCode(code: string) {
-  inject((script) => (script.innerHTML = code))
-})(wm2Polyfill)
